+++ src/components/index.jsx
@@ -4,9 +4,11 @@ import CarouselHome from './homeComponents/carouselHome/carouselHome';
 import FirstSectionNews from './homeComponents/sectionNews/firstSectionNews';
 import SecondSectionNews from './homeComponents/sectionNews/secondSectionNews';
 import FrequentQuestions from './homeComponents/frequentQuestions/frequentQuestions';
+import BannerHome from './homeComponents/bannerHome/bannerHome';
 
 export {
     Navbar,
+    BannerHome,
     CarouselHome,
     FirstSectionNews,
     SecondSectionNews,
